requirements.txt
# Core dependencies
httpx>=0.24.0
aiohttp>=3.8.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API clients
PyGithub>=2.1.0
requests>=2.31.0
requests-cache>=1.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Analysis
scikit-learn>=1.3.0
transformers>=4.30.0
networkx>=3.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
